chore: remove CODEOWNERS
Codeowners are a great idea, but the email is so spammy that it is impossible to see actual emails you need to pay attention to.

I've already disabled required codeowner approval, but this did not stop the emails.  We still require a review, just not from codeowners.  I think we can be big boys and girls and do the right thing without the enforced list of "approved" approvers.