feat: implements video and video gif component
Adds the Video and VideoGif components along with documentation, styling, and tests.
Video is a basic wrapper around the base video HTML tag, however VideoGif is slightly
more complex. It swaps out between a Video and Image component based on if the GIF is
playing or not and adds a hover icon the user can click to start/stop the GIF.